+++ inspect_wandb/__init__.py
@@ -6,4 +6,4 @@ if INSTALLED_EXTRAS["weave"]:
 else:
     __all__ = ["wandb_models_hooks"]
 
-__version__ = "0.1.0"
+__version__ = "0.2.2"