+++ inspect_wandb/config/extras_manager.py
@@ -1,4 +1,4 @@
-import importlib.util
+from importlib.util import find_spec
 
 class ExtrasManager:
     def __init__(self):
@@ -10,13 +10,13 @@ class ExtrasManager:
         return self.extras
 
     def _check_for_weave_extra(self) -> None:
-        if importlib.util.find_spec("weave") is not None:
+        if find_spec("weave") is not None:
             self.extras["weave"] = True
         else:
             self.extras["weave"] = False
 
     def _check_for_viz_extra(self) -> None:
-        if importlib.util.find_spec("inspect_viz") is not None and importlib.util.find_spec("playwright") is not None:
+        if find_spec("inspect_viz") is not None and find_spec("playwright") is not None:
             self.extras["viz"] = False
         else:
             self.extras["viz"] = False