+++ inspect_wandb/config/settings/models.py
@@ -2,20 +2,14 @@ from pydantic import BaseModel, Field, field_validator
 from typing import Any, Literal
 from pydantic_settings import SettingsConfigDict
 from inspect_wandb.config.settings.base import InspectWandBBaseSettings
-import os
+from os import getenv
 
 
 class EnvironmentValidations(BaseModel):
-    """
-    A set of environment variables which should be validated before enabling the integration.
-    """
     wandb_base_url: str | None = Field(default=None, description="The base URL of the wandb instance")
     wandb_api_key: str | None = Field(default=None, description="The API key for the wandb instance")
 
 class ModelsSettings(InspectWandBBaseSettings):
-    """
-    Settings model for the Models integration.
-    """
 
     model_config = SettingsConfigDict(
         env_prefix="INSPECT_WANDB_MODELS_", 
@@ -44,10 +38,10 @@ class ModelsSettings(InspectWandBBaseSettings):
     @classmethod
     def validate_environment_variables(cls, v: EnvironmentValidations | None) -> EnvironmentValidations | None:
         if v is not None:
-            if v.wandb_base_url is not None and (env_wandb_base_url := os.getenv("WANDB_BASE_URL")) != v.wandb_base_url:
+            if v.wandb_base_url is not None and (env_wandb_base_url := getenv("WANDB_BASE_URL")) != v.wandb_base_url:
                 cls.enabled = False
                 raise ValueError(f"WANDB_BASE_URL does not match the value in the environment. Validation URL: {v.wandb_base_url}, Environment URL: {env_wandb_base_url}")
-            if v.wandb_api_key is not None and (env_wandb_api_key := os.getenv("WANDB_API_KEY")) != v.wandb_api_key:
+            if v.wandb_api_key is not None and (env_wandb_api_key := getenv("WANDB_API_KEY")) != v.wandb_api_key:
                 cls.enabled = False
                 raise ValueError(f"WANDB_API_KEY does not match the value in the environment. Validation Key: {v.wandb_api_key}, Environment Key: {env_wandb_api_key}")
         return v