+++ inspect_wandb/viz/inspect_viz_writer.py
@@ -3,21 +3,17 @@ from inspect_viz.plot import write_png_async
 import logging
 from inspect_viz.view.beta import scores_heatmap
 from inspect_viz import Data
-import os
+from pathlib import Path
 import pandas as pd
-import wandb
+from wandb import Run, log, Image
 from inspect_ai.hooks import RunEnd
 from inspect_ai.analysis import evals_df
 
 logger = logging.getLogger(__name__)
 
 class InspectVizWriter:
-    """
-    Class for managing the generation and writing of visualisations with inspect_viz.
-    These visualisations are saved as images to the wandb Models run, if the extra is enabled.
-    """
 
-    async def log_scores_heatmap(self, data: RunEnd, run: wandb.Run) -> None:
+    async def log_scores_heatmap(self, data: RunEnd, run: Run) -> None:
         try:
             logs = [log.location for log in data.logs]
             run.config["logs"] = logs
@@ -33,8 +29,7 @@ class InspectVizWriter:
         await self._log_image(data.run_id, plot, "scores_heatmap")
 
     async def _log_image(self, run_id: str, plot: Component, name: str) -> None:
-        path = f"./.plots/{run_id}/{name}.png"
-        if not os.path.exists(f"./.plots/{run_id}"):
-            os.makedirs(f"./.plots/{run_id}")
-        await write_png_async(path, plot)
-        wandb.log({name: wandb.Image(path)})
+        path = Path("./.plots") / run_id / f"{name}.png"
+        path.parent.mkdir(parents=True, exist_ok=True)
+        await write_png_async(str(path), plot)
+        log({name: Image(str(path))})